# backend/openapi.yaml
openapi: "3.1.0"
info:
  title: YakinLulus.id API
  version: "1.0.0"
  description: |
    Computer-Based Testing (CBT) platform API for YakinLulus.id.
    Authentication via Bearer JWT. All authenticated routes use
    the `Authorization: Bearer <token>` header.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
paths:
...
  /questions/{id}:
    get:
      operationId: getQuestion
      summary: Get question by ID
      tags: [Question Bank]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Question data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
    put:
      operationId: updateQuestion
      summary: Update question (admin/staff/teacher)
      tags: [Question Bank]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stem: { type: string }
                type: { type: string }
                difficulty: { type: string }
                explanation: { type: string }
                points: { type: integer }
      responses:
        "200":
          description: Question updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
    delete:
      operationId: deleteQuestion
      summary: Delete question (admin/staff/teacher)
      tags: [Question Bank]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Question deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
...
  /exams/{id}:
    get:
      operationId: getExam
      summary: Get exam by ID
      tags: [CBT Engine]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Exam data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
    put:
      operationId: updateExam
      summary: Update exam (admin/staff/teacher)
      tags: [CBT Engine]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                duration_minutes: { type: integer }
                passing_score: { type: integer }
      responses:
        "200":
          description: Exam updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
    delete:
      operationId: deleteExam
      summary: Delete exam (admin/staff/teacher)
      tags: [CBT Engine]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Exam deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
...
  /materials/{id}:
    get:
      operationId: getMaterial
      summary: Get material by ID
      tags: [Material]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Material data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
    put:
      operationId: updateMaterial
      summary: Update material (admin/staff/teacher)
      tags: [Material]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                content: { type: string }
                type: { type: string }
                order: { type: integer }
      responses:
        "200":
          description: Material updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
    delete:
      operationId: deleteMaterial
      summary: Delete material (admin/staff/teacher)
      tags: [Material]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Material deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
    Success:
      type: object
      properties:
        success:
          type: boolean
          description: Operation result
          example: true
        message:
          type: string
          description: Human-readable message
          example: Operation successful
        data:
          description: Response payload
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp
...
security:
  - BearerAuth: []